insert, delete, and edit methods vacancies and users are done

+++ src/main/java/com/example/demo/controller/ResumeController.java
@@ -47,7 +47,7 @@ public class ResumeController {
     }
 
     @DeleteMapping("/{id}")
-    public HttpStatus deleteResume(@PathVariable("id") int id) {
+    public HttpStatus deleteResume(@PathVariable("id") long id) {
         resumeService.deleteById(id);
         return HttpStatus.OK;
     }
