+++ src/main/java/com/example/demo/model/Vacancy.java
@@ -17,8 +17,8 @@ public class Vacancy {
     private Integer expFrom;
     private Integer expTo;
     private boolean isActive;
-    private Integer userId;
+    private Long userId;
     private LocalDateTime createdDate;
     private LocalDateTime updateTime;
-    private Integer id;
+    private Long id;
 }
